agents/python-agent/requirements.txt
Pillow==11.1.0
websockets==15.0.1
